+++ src/controller_examples/zookeeper_controller/exec/zookeeper_api.rs
@@ -3,7 +3,6 @@
 use crate::external_api::exec::*;
 use crate::pervasive_ext::to_view::*;
 use crate::zookeeper_controller::common::*;
-use crate::zookeeper_controller::exec::zookeepercluster::*;
 use crate::zookeeper_controller::spec::zookeeper_api::{
     ZKAPIInputView, ZKAPIOutputView, ZKNodeResultView,
 };
