+++ src/controller_examples/zookeeper_controller/spec/mod.rs
@@ -1,5 +1,5 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 pub mod reconciler;
+pub mod types;
 pub mod zookeeper_api;
-pub mod zookeepercluster;
